requirements.txt
django-heroku
apparmor==2.13.4
appdirs==1.4.4
application-utility==1.3.2
asgiref==3.2.8
btrfsutil==1.2.0
CacheControl==0.12.6
catfish==1.4.13
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0.post1
distlib==0.3.0
distro==1.5.0
Django==3.0.7
docopt==0.6.2
gufw==20.4.0
html5lib==1.0.1
idna==2.9
keyutils==0.6
lensfun==0.3.95
LibAppArmor==2.13.4
lightdm-gtk-greeter-settings==1.2.2
louis==3.14.0
menulibre==2.2.1
msgpack==1.0.0
mugshot==0.4.2
npyscreen==4.10.5
ordered-set==3.1.1
packaging==20.4
pacman-mirrors==4.16.4
pep517==0.8.2
pexpect==4.8.0
Pillow==6.2.1
pipupgrade==1.7.2
progress==1.5
psutil==5.7.0
ptyprocess==0.6.0
pulsemixer==1.5.1
pwquality==1.4.2
pycairo==1.19.1
PyGObject==3.36.1
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.8.0
pysmbc==1.0.22
python-distutils-extra==2.39
pytoml==0.1.21
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
rados==2.0.0
rbd==2.0.0
reportlab==3.5.42
requests==2.23.0
retrying==1.3.3
rgw==2.0.0
sip==4.19.22
six==1.15.0
sqlparse==0.3.1
team==1.0
toml==0.10.1
udiskie==2.2.0
ufw==0.36
urllib3==1.25.9
wakatime==13.0.7
webencodings==0.5.1
psycopg2==2.8.5
gunicorn
pygithub